requirements.txt
requests>=2.31.0
requests-oauthlib>=1.3.1
python-dotenv>=1.0.0
click>=8.0.0
pyyaml>=6.0.0
rich>=13.0.0
tabulate>=0.9.0